Don't spawn a new metrics loop updating the outbox state for each inbox
